refactor: move cluster-scoped MR controllers to internal/controller/cluster
