+++ internal/controller/cloudian.go
@@ -20,28 +20,14 @@ import (
 	"github.com/crossplane/crossplane-runtime/v2/pkg/controller"
 	ctrl "sigs.k8s.io/controller-runtime"
 
-	"github.com/statnett/provider-cloudian/internal/controller/accesskey"
-	"github.com/statnett/provider-cloudian/internal/controller/config"
-	"github.com/statnett/provider-cloudian/internal/controller/group"
-	"github.com/statnett/provider-cloudian/internal/controller/groupqualityofservicelimits"
-	"github.com/statnett/provider-cloudian/internal/controller/user"
-	"github.com/statnett/provider-cloudian/internal/controller/userqualityofservicelimits"
+	controllercluster "github.com/statnett/provider-cloudian/internal/controller/cluster"
 )
 
 // Setup creates all Cloudian controllers with the supplied logger and adds them to
 // the supplied manager.
 func Setup(mgr ctrl.Manager, o controller.Options) error {
-	for _, setup := range []func(ctrl.Manager, controller.Options) error{
-		accesskey.Setup,
-		config.Setup,
-		group.Setup,
-		groupqualityofservicelimits.Setup,
-		user.Setup,
-		userqualityofservicelimits.Setup,
-	} {
-		if err := setup(mgr, o); err != nil {
-			return err
-		}
+	if err := controllercluster.Setup(mgr, o); err != nil {
+		return err
 	}
 	return nil
 }
