+++ internal/controller/cluster/cluster.go
@@ -1,0 +1,47 @@
+/*
+Copyright 2020 The Crossplane Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package cluster
+
+import (
+	"github.com/crossplane/crossplane-runtime/v2/pkg/controller"
+	ctrl "sigs.k8s.io/controller-runtime"
+
+	"github.com/statnett/provider-cloudian/internal/controller/cluster/accesskey"
+	"github.com/statnett/provider-cloudian/internal/controller/cluster/config"
+	"github.com/statnett/provider-cloudian/internal/controller/cluster/group"
+	"github.com/statnett/provider-cloudian/internal/controller/cluster/groupqualityofservicelimits"
+	"github.com/statnett/provider-cloudian/internal/controller/cluster/user"
+	"github.com/statnett/provider-cloudian/internal/controller/cluster/userqualityofservicelimits"
+)
+
+// Setup creates all Cloudian controllers related to cluster scoped MRs
+// with the supplied logger and adds them to the supplied manager.
+func Setup(mgr ctrl.Manager, o controller.Options) error {
+	for _, setup := range []func(ctrl.Manager, controller.Options) error{
+		accesskey.Setup,
+		config.Setup,
+		group.Setup,
+		groupqualityofservicelimits.Setup,
+		user.Setup,
+		userqualityofservicelimits.Setup,
+	} {
+		if err := setup(mgr, o); err != nil {
+			return err
+		}
+	}
+	return nil
+}
